refactor: return type of searchStreams
