+++ packages/client/test/unit/searchStreams.test.ts
@@ -2,7 +2,7 @@ import 'reflect-metadata'
 import { BigNumber } from '@ethersproject/bignumber'
 import { StreamID, toStreamID } from 'streamr-client-protocol'
 import { searchStreams, SearchStreamsResultItem } from '../../src/registry/searchStreams'
-import { collect } from '../../src/utils/GeneratorUtils'
+import { collect } from '../../src/utils/iterators'
 import { Stream } from '../../src/Stream'
 import { SynchronizedGraphQLClient } from '../../src/utils/SynchronizedGraphQLClient'
 import { randomEthereumAddress } from 'streamr-test-utils'
@@ -31,7 +31,7 @@ const createMockGraphQLClient = (resultItems: SearchStreamsResultItem[]): Pick<S
         fetchPaginatedResults: async function* () {
             yield* resultItems
         } as any
-    } 
+    }
 }
 
 describe('searchStreams', () => {
